src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
kombu==5.3.7
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
